backend/requirements.txt
# Truth in Listings - Backend Dependencies
# Python 3.9+

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Image processing
Pillow==10.2.0

# Geospatial
geopy==2.4.1

# Evaluation (for evaluation module)
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1

# CORS
python-cors==1.0.0
